+++ src/signCredentialDeployment.c
@@ -13,7 +13,6 @@
 #include "util.h"
 
 static signCredentialDeploymentContext_t *ctx = &global.signCredentialDeploymentContext;
-static cx_sha256_t attributeHash;
 static tx_state_t *tx_state = &global_tx_state;
 
 void processNextVerificationKey(void);
@@ -24,9 +23,9 @@ void handleSignCredentialDeployment(
     volatile unsigned int *flags,
     bool isInitialCall);
 
-UX_STEP_CB(ux_credential_deployment_initial_flow_0_step, nn, sendSuccessNoIdle(), {"Review", "details"});
+UX_STEP_CB(ux_credential_deployment_review_details, nn, sendSuccessNoIdle(), {"Review", "details"});
+
 UX_STEP_CB(ux_update_credentials_initial_flow_1_step, nn, sendSuccessNoIdle(), {"Continue", "with transaction"});
-UX_FLOW(ux_credential_deployment_initial_flow, &ux_credential_deployment_initial_flow_0_step);
 
 UX_FLOW(
     ux_update_credentials_initial_flow,
@@ -34,12 +33,26 @@ UX_FLOW(
     &ux_sign_flow_account_sender_view,
     &ux_update_credentials_initial_flow_1_step);
 
-UX_STEP_CB(
+UX_STEP_NOCB(
     ux_credential_deployment_verification_key_flow_0_step,
     bnnn_paging,
-    processNextVerificationKey(),
     {.title = "Public key", .text = (char *) global.signCredentialDeploymentContext.accountVerificationKey});
-UX_FLOW(ux_credential_deployment_verification_key_flow, &ux_credential_deployment_verification_key_flow_0_step);
+
+UX_STEP_CB(
+    ux_credential_deployment_verification_key_flow_1_step,
+    nn,
+    processNextVerificationKey(),
+    {"Continue", "with transaction"});
+
+UX_FLOW(
+    ux_credential_deployment_verification_key_flow,
+    &ux_credential_deployment_verification_key_flow_0_step,
+    &ux_credential_deployment_verification_key_flow_1_step);
+UX_FLOW(
+    ux_credential_deployment_verification_key_flow_with_intro,
+    &ux_credential_deployment_review_details,
+    &ux_credential_deployment_verification_key_flow_0_step,
+    &ux_credential_deployment_verification_key_flow_1_step);
 
 UX_STEP_NOCB(
     ux_credential_deployment_threshold_flow_0_step,
@@ -50,10 +63,6 @@ UX_STEP_CB(
     bn,
     sendSuccessNoIdle(),
     {"AR threshold", (char *) global.signCredentialDeploymentContext.anonymityRevocationThreshold});
-UX_FLOW(
-    ux_credential_deployment_threshold_flow,
-    &ux_credential_deployment_threshold_flow_0_step,
-    &ux_credential_deployment_threshold_flow_1_step);
 
 UX_STEP_NOCB(
     ux_sign_credential_deployment_0_step,
@@ -71,13 +80,38 @@ UX_STEP_CB(
     {&C_icon_crossmark, "Decline to", "sign details"});
 
 UX_FLOW(
-    ux_sign_credential_deployment_existing,
+    ux_sign_credential_deployment_existing_with_intro,
+    &ux_credential_deployment_review_details,
+    &ux_credential_deployment_verification_key_flow_0_step,
+    &ux_credential_deployment_threshold_flow_0_step,
+    &ux_credential_deployment_threshold_flow_1_step,
     &ux_sign_credential_deployment_0_step,
     &ux_sign_credential_deployment_1_step,
     &ux_sign_credential_deployment_2_step);
 
 UX_FLOW(
+    ux_sign_credential_deployment_existing,
+    &ux_credential_deployment_verification_key_flow_0_step,
+    &ux_credential_deployment_threshold_flow_0_step,
+    &ux_credential_deployment_threshold_flow_1_step,
+    &ux_sign_credential_deployment_0_step,
+    &ux_sign_credential_deployment_1_step,
+    &ux_sign_credential_deployment_2_step);
+
+UX_FLOW(
+    ux_sign_credential_deployment_new_with_intro,
+    &ux_credential_deployment_review_details,
+    &ux_credential_deployment_verification_key_flow_0_step,
+    &ux_credential_deployment_threshold_flow_0_step,
+    &ux_credential_deployment_threshold_flow_1_step,
+    &ux_sign_credential_deployment_1_step,
+    &ux_sign_credential_deployment_2_step);
+
+UX_FLOW(
     ux_sign_credential_deployment_new,
+    &ux_credential_deployment_verification_key_flow_0_step,
+    &ux_credential_deployment_threshold_flow_0_step,
+    &ux_credential_deployment_threshold_flow_1_step,
     &ux_sign_credential_deployment_1_step,
     &ux_sign_credential_deployment_2_step);
 
@@ -88,10 +122,6 @@ UX_STEP_CB(
     {.title = "Rem. credential", .text = (char *) global.signCredentialDeploymentContext.credentialId});
 UX_FLOW(ux_sign_credential_update_id, &ux_sign_credential_update_id_0_step);
 
-/*
- * The UI flow for the final part of the update credential, which displays
- * the threshold and allows the user to either sign or decline.
- */
 UX_STEP_NOCB(
     ux_sign_credential_update_threshold_0_step,
     bnnn_paging,
@@ -123,16 +153,16 @@ void processNextVerificationKey(void) {
 
 void parseVerificationKey(uint8_t *buffer) {
     // Hash key index
-    cx_hash((cx_hash_t *) &tx_state->hash, 0, buffer, 1, NULL, 0);
+    updateHash((cx_hash_t *) &tx_state->hash, buffer, 1);
     buffer += 1;
 
     // Hash schemeId
-    cx_hash((cx_hash_t *) &tx_state->hash, 0, buffer, 1, NULL, 0);
+    updateHash((cx_hash_t *) &tx_state->hash, buffer, 1);
     buffer += 1;
 
     uint8_t verificationKey[32];
     memmove(verificationKey, buffer, 32);
-    cx_hash((cx_hash_t *) &tx_state->hash, 0, verificationKey, 32, NULL, 0);
+    updateHash((cx_hash_t *) &tx_state->hash, verificationKey, 32);
 
     // Convert to a human-readable format.
     toPaginatedHex(
@@ -141,9 +171,6 @@ void parseVerificationKey(uint8_t *buffer) {
         ctx->accountVerificationKey,
         sizeof(ctx->accountVerificationKey));
     ctx->numberOfVerificationKeys -= 1;
-
-    // Show to the user.
-    ux_flow_init(0, ux_credential_deployment_verification_key_flow, NULL);
 }
 
 // APDU parameters specific to credential deployment transaction (multiple packets protocol).
@@ -189,7 +216,7 @@ void handleSignUpdateCredential(
         dataBuffer += hashAccountTransactionHeaderAndKind(dataBuffer, UPDATE_CREDENTIALS);
 
         ctx->credentialDeploymentCount = dataBuffer[0];
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 1);
         if (ctx->credentialDeploymentCount == 0) {
             ctx->updateCredentialState = TX_UPDATE_CREDENTIAL_ID_COUNT;
         } else {
@@ -202,7 +229,7 @@ void handleSignUpdateCredential(
         p2 == P2_CREDENTIAL_CREDENTIAL_INDEX && ctx->updateCredentialState == TX_UPDATE_CREDENTIAL_CREDENTIAL_INDEX &&
         ctx->credentialDeploymentCount > 0) {
         // Add the credential index to the hash
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 1);
         ctx->updateCredentialState = TX_UPDATE_CREDENTIAL_CREDENTIAL;
         sendSuccessNoIdle();
     } else if (
@@ -211,7 +238,7 @@ void handleSignUpdateCredential(
         handleSignCredentialDeployment(dataBuffer, p1, p2, flags, false);
     } else if (p2 == P2_CREDENTIAL_ID_COUNT && ctx->updateCredentialState == TX_UPDATE_CREDENTIAL_ID_COUNT) {
         ctx->credentialIdCount = dataBuffer[0];
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 1);
 
         if (ctx->credentialIdCount == 0) {
             ctx->updateCredentialState = TX_UPDATE_CREDENTIAL_THRESHOLD;
@@ -220,7 +247,7 @@ void handleSignUpdateCredential(
         }
         sendSuccessNoIdle();
     } else if (p2 == P2_CREDENTIAL_ID && ctx->updateCredentialState == TX_UPDATE_CREDENTIAL_ID) {
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 48, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 48);
         toPaginatedHex(dataBuffer, 48, ctx->credentialId, sizeof(ctx->credentialId));
 
         ctx->credentialIdCount -= 1;
@@ -233,7 +260,7 @@ void handleSignUpdateCredential(
     } else if (p2 == P2_THRESHOLD && ctx->updateCredentialState == TX_UPDATE_CREDENTIAL_THRESHOLD) {
         uint8_t threshold = dataBuffer[0];
         bin2dec(ctx->threshold, sizeof(ctx->threshold), threshold);
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 1);
 
         ux_flow_init(0, ux_sign_credential_update_threshold, NULL);
         *flags |= IO_ASYNCH_REPLY;
@@ -258,11 +285,12 @@ void handleSignCredentialDeployment(
         // Initialize values.
         cx_sha256_init(&tx_state->hash);
         ctx->state = TX_CREDENTIAL_DEPLOYMENT_VERIFICATION_KEYS_LENGTH;
+        ctx->showIntro = true;
 
-        ux_flow_init(0, ux_credential_deployment_initial_flow, NULL);
+        sendSuccessNoIdle();
     } else if (p1 == P1_VERIFICATION_KEY_LENGTH && ctx->state == TX_CREDENTIAL_DEPLOYMENT_VERIFICATION_KEYS_LENGTH) {
         ctx->numberOfVerificationKeys = dataBuffer[0];
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 1);
         ctx->state = TX_CREDENTIAL_DEPLOYMENT_VERIFICATION_KEY;
         sendSuccessNoIdle();
     } else if (p1 == P1_VERIFICATION_KEY) {
@@ -271,6 +299,23 @@ void handleSignCredentialDeployment(
         } else {
             THROW(ERROR_INVALID_STATE);
         }
+
+        if (ctx->numberOfVerificationKeys > 0) {
+            if (ctx->showIntro) {
+                // For the first key we also display the initial view.
+                ctx->showIntro = false;
+                ux_flow_init(0, ux_credential_deployment_verification_key_flow_with_intro, NULL);
+            } else {
+                // Display a key with continue here.
+                ux_flow_init(0, ux_credential_deployment_verification_key_flow, NULL);
+            }
+            *flags |= IO_ASYNCH_REPLY;
+        } else {
+            // Do not display the last verification key here. This is deferred to the final UI flow.
+            ctx->state = TX_CREDENTIAL_DEPLOYMENT_SIGNATURE_THRESHOLD;
+            sendSuccessNoIdle();
+        }
+
     } else if (p1 == P1_SIGNATURE_THRESHOLD && ctx->state == TX_CREDENTIAL_DEPLOYMENT_SIGNATURE_THRESHOLD) {
         if (ctx->numberOfVerificationKeys != 0) {
             THROW(
@@ -279,17 +324,17 @@ void handleSignCredentialDeployment(
 
         // Parse signature threshold.
         bin2dec(ctx->signatureThreshold, sizeof(ctx->signatureThreshold), dataBuffer[0]);
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 1);
         dataBuffer += 1;
 
         // Parse the RegIdCred, but do not display it, as the user cannot feasibly verify it.
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 48, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 48);
         dataBuffer += 48;
 
         // Parse identity provider index.
         // We do not show the identity provider id, because it is infeasible for the user to validate it,
         // and there are no known reasonable attacks made possible by replacing this.
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 4, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 4);
         dataBuffer += 4;
 
         // Parse anonymity revocation threshold.
@@ -297,25 +342,21 @@ void handleSignCredentialDeployment(
             numberToText(ctx->anonymityRevocationThreshold, sizeof(ctx->anonymityRevocationThreshold), dataBuffer[0]);
         memmove(ctx->anonymityRevocationThreshold + offset, " out of ", 8);
         offset += 8;
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 1);
         dataBuffer += 1;
 
         // Parse the length of the following list of anonymity revokers.
         ctx->anonymityRevocationListLength = U2BE(dataBuffer, 0);
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 2, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 2);
         // Add the total amount of revokers to the display of threshold to get "x out of y"
         bin2dec(
             ctx->anonymityRevocationThreshold + offset,
             sizeof(ctx->anonymityRevocationThreshold) - offset,
             ctx->anonymityRevocationListLength);
 
-        // Initialize values for later.
-        cx_sha256_init(&attributeHash);
-
         ctx->state = TX_CREDENTIAL_DEPLOYMENT_AR_IDENTITY;
 
-        // Display the loaded data.
-        ux_flow_init(0, ux_credential_deployment_threshold_flow, NULL);
+        sendSuccessNoIdle();
     } else if (p1 == P1_AR_IDENTITY && ctx->state == TX_CREDENTIAL_DEPLOYMENT_AR_IDENTITY) {
         if (ctx->anonymityRevocationListLength == 0) {
             // Invalid state, sender says ar identity pair is incoming, but we already received all.
@@ -325,7 +366,7 @@ void handleSignCredentialDeployment(
         // Parse ArIdentity
         // We do not show the AR identity id, because it is infeasible for the user to validate it,
         // and there are no known reasonable attacks made possible by replacing this.
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 4, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 4);
         dataBuffer += 4;
 
         // Parse enc_id_cred_pub_share
@@ -333,12 +374,7 @@ void handleSignCredentialDeployment(
         // to validate.
         uint8_t encIdCredPubShare[96];
         memmove(encIdCredPubShare, dataBuffer, 96);
-        toPaginatedHex(
-            encIdCredPubShare,
-            sizeof(encIdCredPubShare),
-            ctx->encIdCredPubShare,
-            sizeof(ctx->encIdCredPubShare));
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, encIdCredPubShare, 96, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, encIdCredPubShare, 96);
 
         if (ctx->anonymityRevocationListLength == 1) {
             ctx->state = TX_CREDENTIAL_DEPLOYMENT_CREDENTIAL_DATES;
@@ -348,12 +384,12 @@ void handleSignCredentialDeployment(
     } else if (p1 == P1_CREDENTIAL_DATES && ctx->state == TX_CREDENTIAL_DEPLOYMENT_CREDENTIAL_DATES) {
         // hash valid to and created at
         // We don't show these values, because only the dates on the identity object can be accepted by the chain.
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 6, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 6);
         dataBuffer += 6;
 
         // Read attribute list length
         ctx->attributeListLength = U2BE(dataBuffer, 0);
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 2, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 2);
 
         if (ctx->attributeListLength == 0) {
             ctx->state = TX_CREDENTIAL_DEPLOYMENT_LENGTH_OF_PROOFS;
@@ -371,34 +407,24 @@ void handleSignCredentialDeployment(
         uint8_t attributeTag[1];
         memmove(attributeTag, dataBuffer, 1);
         dataBuffer += 1;
-        cx_hash((cx_hash_t *) &attributeHash, 0, attributeTag, 1, NULL, 0);
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, attributeTag, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, attributeTag, 1);
 
         // Parse attribute length, so we know how much to parse in next packet.
         uint8_t attributeValueLength[1];
         memmove(attributeValueLength, dataBuffer, 1);
         ctx->attributeValueLength = attributeValueLength[0];
-        cx_hash((cx_hash_t *) &attributeHash, 0, attributeValueLength, 1, NULL, 0);
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, attributeValueLength, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, attributeValueLength, 1);
 
         ctx->state = TX_CREDENTIAL_DEPLOYMENT_ATTRIBUTE_VALUE;
         // Ask computer for the attribute value.
         sendSuccessNoIdle();
     } else if (p1 == P1_ATTRIBUTE_VALUE && ctx->state == TX_CREDENTIAL_DEPLOYMENT_ATTRIBUTE_VALUE) {
         // Add attribute value to the hash.
-        cx_hash((cx_hash_t *) &attributeHash, 0, dataBuffer, ctx->attributeValueLength, NULL, 0);
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, ctx->attributeValueLength, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, ctx->attributeValueLength);
         ctx->attributeListLength -= 1;
 
         // We have processed all attributes
         if (ctx->attributeListLength == 0) {
-            uint8_t attributeHashBytes[32];
-            cx_hash((cx_hash_t *) &attributeHash, CX_LAST, NULL, 0, attributeHashBytes, 32);
-            toPaginatedHex(
-                attributeHashBytes,
-                sizeof(attributeHashBytes),
-                ctx->attributeHashDisplay,
-                sizeof(ctx->attributeHashDisplay));
             ctx->state = TX_CREDENTIAL_DEPLOYMENT_LENGTH_OF_PROOFS;
             sendSuccessNoIdle();
         } else {
@@ -409,17 +435,17 @@ void handleSignCredentialDeployment(
     } else if (p1 == P1_LENGTH_OF_PROOFS && ctx->state == TX_CREDENTIAL_DEPLOYMENT_LENGTH_OF_PROOFS) {
         ctx->proofLength = U4BE(dataBuffer, 0);
         if (p2 == P2_CREDENTIAL_CREDENTIAL) {
-            cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 4, NULL, 0);
+            updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 4);
         }
         ctx->state = TX_CREDENTIAL_DEPLOYMENT_PROOFS;
         sendSuccessNoIdle();
     } else if (p1 == P1_PROOFS && ctx->state == TX_CREDENTIAL_DEPLOYMENT_PROOFS) {
         if (ctx->proofLength > MAX_CDATA_LENGTH) {
-            cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, MAX_CDATA_LENGTH, NULL, 0);
+            updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, MAX_CDATA_LENGTH);
             ctx->proofLength -= MAX_CDATA_LENGTH;
             sendSuccessNoIdle();
         } else {
-            cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, ctx->proofLength, NULL, 0);
+            updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, ctx->proofLength);
 
             // If an update credential transaction, then update state to next step.
             if (p2 == P2_CREDENTIAL_CREDENTIAL && ctx->updateCredentialState == TX_UPDATE_CREDENTIAL_CREDENTIAL &&
@@ -440,12 +466,16 @@ void handleSignCredentialDeployment(
     } else if (p1 == P1_NEW_OR_EXISTING && ctx->state == TX_CREDENTIAL_DEPLOYMENT_NEW_OR_EXISTING) {
         // 0 indicates new, 1 indicates existing
         uint8_t newOrExisting = dataBuffer[0];
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 1);
         dataBuffer += 1;
 
         if (newOrExisting == 0) {
-            cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 8, NULL, 0);
-            ux_flow_init(0, ux_sign_credential_deployment_new, NULL);
+            updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 8);
+            if (ctx->showIntro) {
+                ux_flow_init(0, ux_sign_credential_deployment_new_with_intro, NULL);
+            } else {
+                ux_flow_init(0, ux_sign_credential_deployment_new, NULL);
+            }
         } else if (newOrExisting == 1) {
             uint8_t accountAddress[32];
             memmove(accountAddress, dataBuffer, 32);
@@ -457,9 +487,13 @@ void handleSignCredentialDeployment(
                 THROW(ERROR_INVALID_TRANSACTION);
             }
             ctx->accountAddress[55] = '\0';
+            updateHash((cx_hash_t *) &tx_state->hash, dataBuffer, 32);
 
-            cx_hash((cx_hash_t *) &tx_state->hash, 0, dataBuffer, 32, NULL, 0);
-            ux_flow_init(0, ux_sign_credential_deployment_existing, NULL);
+            if (ctx->showIntro) {
+                ux_flow_init(0, ux_sign_credential_deployment_existing_with_intro, NULL);
+            } else {
+                ux_flow_init(0, ux_sign_credential_deployment_existing, NULL);
+            }
         } else {
             THROW(ERROR_INVALID_TRANSACTION);
         }