+++ src/signTransferWithSchedule.c
@@ -75,7 +75,7 @@ void processNextScheduledAmount(uint8_t *buffer) {
         // The current packet still has additional timestamp/amount pairs to be added to the hash and
         // displayed for the user.
         uint64_t timestamp = U8BE(ctx->buffer, ctx->pos) / 1000;
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, buffer + ctx->pos, 8, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, buffer + ctx->pos, 8);
         ctx->pos += 8;
         int valid = secondsToTm(timestamp, &ctx->time);
         if (valid != 0) {
@@ -90,7 +90,7 @@ void processNextScheduledAmount(uint8_t *buffer) {
         timeToDisplayText(ctx->time, ctx->displayTimestamp, sizeof(ctx->displayTimestamp));
 
         uint64_t amount = U8BE(ctx->buffer, ctx->pos);
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, buffer + ctx->pos, 8, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, buffer + ctx->pos, 8);
         ctx->pos += 8;
         amountToGtuDisplay(ctx->displayAmount, sizeof(ctx->displayAmount), amount);
 
@@ -137,7 +137,7 @@ void handleTransferPairs(uint8_t *cdata, volatile unsigned int *flags) {
 #define P1_MEMO                     0x03
 
 void finishMemoScheduled(volatile unsigned int *flags) {
-    cx_hash((cx_hash_t *) &tx_state->hash, 0, &ctx->remainingNumberOfScheduledAmounts, 1, NULL, 0);
+    updateHash((cx_hash_t *) &tx_state->hash, &ctx->remainingNumberOfScheduledAmounts, 1);
     ctx->state = TX_TRANSFER_WITH_SCHEDULE_TRANSFER_PAIRS;
     startInitialScheduledTransferDisplay(true);
     *flags |= IO_ASYNCH_REPLY;
@@ -167,13 +167,13 @@ void handleSignTransferWithScheduleAndMemo(
             THROW(ERROR_INVALID_PARAM);
         }
 
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 2, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, cdata, 2);
 
         // Update the state to expect the next message to contain the first bytes of the memo.
         ctx->state = TX_TRANSFER_WITH_SCHEDULE_MEMO_START;
         sendSuccessNoIdle();
     } else if (p1 == P1_MEMO && ctx->state == TX_TRANSFER_WITH_SCHEDULE_MEMO_START) {
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, dataLength, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, cdata, dataLength);
 
         // Read initial part of memo and then display it:
         readCborInitial(cdata, dataLength);
@@ -185,7 +185,7 @@ void handleSignTransferWithScheduleAndMemo(
             sendSuccessNoIdle();
         }
     } else if (p1 == P1_MEMO && ctx->state == TX_TRANSFER_WITH_SCHEDULE_MEMO) {
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, dataLength, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, cdata, dataLength);
 
         // Read current part of memo and then display it:
         readCborContent(cdata, dataLength);
@@ -214,7 +214,7 @@ void handleSignTransferWithSchedule(uint8_t *cdata, uint8_t p1, volatile unsigne
         // Store the number of scheduled amounts we are going to receive next.
         ctx->remainingNumberOfScheduledAmounts = cdata[0];
         // Hash schedule length
-        cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 1, NULL, 0);
+        updateHash((cx_hash_t *) &tx_state->hash, cdata, 1);
 
         // Update the state to expect the next message to contain transfer pairs.
         ctx->state = TX_TRANSFER_WITH_SCHEDULE_TRANSFER_PAIRS;