+++ tests/src/credentialDeployment.test.ts
@@ -1,6 +1,6 @@
 import Transport from '@ledgerhq/hw-transport';
 import Zemu from '@zondax/zemu';
-import chunkBuffer from './helpers';
+import chunkBuffer, { toHex } from './helpers';
 import { setupZemu } from './options';
 
 enum CredentialDeploymentType {
@@ -9,28 +9,52 @@ enum CredentialDeploymentType {
 }
 
 /**
+ * Handles (sends data and performs UI actions) the public key part
+ * of the credential deployment transaction.
+ */
+async function handleCredentialDeploymentKeys(sim: Zemu, transport: Transport, ins: number, p2: number, numberOfKeys: number, navigateKeys?: () => Promise<void>) {
+    let data = Buffer.from(toHex(numberOfKeys), 'hex');
+    await transport.send(0xe0, ins, 0x0A, p2, data);
+    let showIntro = true;
+
+    // Send all keys except the final key and handle navigating the UI for the keys.
+    // This is only relevant if there are multiple keys, as the UI for the final key
+    // is deferred to the end of the flow.
+    for (let i = 0; i < numberOfKeys - 1; i++) {
+        data = Buffer.from(toHex(i + 1) + '00f78929ec8a9819f6ae2e10e79522b6b311949635fecc3d924d9d1e23f8e9e1c3', 'hex');
+        transport.send(0xe0, ins, 0x01, p2, data);
+
+        if (showIntro) {
+            showIntro = false;
+            await sim.clickRight(undefined, true);
+        }
+
+        if (navigateKeys) {
+            await navigateKeys();
+        } else {
+            throw new Error('Navigate keys should be defined when sending multiple keys.');
+        }
+    }
+
+    data = Buffer.from('0000f78929ec8a9819f6ae2e10e79522b6b311949635fecc3d924d9d1e23f8e9e1c3', 'hex');
+    await transport.send(0xe0, ins, 0x01, p2, data);
+}
+
+/**
  * Handles the shared part of credential deployment / update credentials.
  */
 async function sharedCredentialDeployment(
-    snapshot: any,
-    sim: Zemu, transport: Transport,
+    sim: Zemu,
+    transport: Transport,
     ins: number,
     p2: number,
-    handleKeyUi: () => Promise<any>,
+    numberOfKeys?: number,
+    navigateKeys?: () => Promise<void>
 ) {
-    let data = Buffer.from('01', 'hex');
-    await transport.send(0xe0, ins, 0x0A, p2, data);
+    await handleCredentialDeploymentKeys(sim, transport, ins, p2, numberOfKeys ? numberOfKeys : 1, navigateKeys);
 
-    data = Buffer.from('0000f78929ec8a9819f6ae2e10e79522b6b311949635fecc3d924d9d1e23f8e9e1c3', 'hex');
-    transport.send(0xe0, ins, 0x01, p2, data);
-    await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
-    const snapshot1 = await handleKeyUi();
-
-    data = Buffer.from('ff85d8a7aa296c162e4e2f0d6bfbdc562db240e28942f7f3ddef6979a1133b5c719ec3581869aaf88388824b0f6755e63c0000f013010001', 'hex');
-    transport.send(0xe0, ins, 0x02, p2, data);
-    await sim.waitUntilScreenIsNot(snapshot1);
-    await sim.clickRight();
-    const snapshot2 = await sim.clickBoth(undefined, false);
+    let data = Buffer.from('ff85d8a7aa296c162e4e2f0d6bfbdc562db240e28942f7f3ddef6979a1133b5c719ec3581869aaf88388824b0f6755e63c0000f013010001', 'hex');
+    await transport.send(0xe0, ins, 0x02, p2, data);
 
     data = Buffer.from('000f0301aca024ce6083d4956edad825c3721da9b61e5b3712606ba1465f7818a43849121bdb3e4d99624e9a74b9436cc8948d178b9b144122aa070372e3fadee4998e1cc21161186a3d19698ad245e10912810df1aaddda16a27f654716108e27758099', 'hex');
     await transport.send(0xe0, ins, 0x03, p2, data);
@@ -52,35 +76,33 @@ async function sharedCredentialDeployment(
 
     for (const chunk of chunkedProofs) {
         await transport.send(0xe0, ins, 0x08, p2, chunk);
-    }
-
-    return snapshot2;
+    };
 }
 
 async function credentialDeployment(
     sim: Zemu, transport: Transport, type: CredentialDeploymentType, expectedSignature: string,
     handleKeyUi: () => Promise<any>,
     handleAddressUi: () => Promise<void>,
+    numberOfKeys?: number,
+    navigateKeys?: () => Promise<void>
 ) {
     let data = Buffer.from('080000045100000000000000000000000000000000000000020000000000000000', 'hex');
-    transport.send(0xe0, 0x04, 0x00, 0x00, data);
-    await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
-    const snapshot1 = await sim.clickBoth(undefined, false);
+    await transport.send(0xe0, 0x04, 0x00, 0x00, data);
 
-    const snapshot2 = await sharedCredentialDeployment(
-        snapshot1, sim, transport, 0x04, 0x00, handleKeyUi,
+    await sharedCredentialDeployment(
+        sim, transport, 0x04, 0x00, numberOfKeys, navigateKeys
     );
 
     let tx;
     if (type === CredentialDeploymentType.NEW) {
         data = Buffer.from('00000000006040F27E', 'hex');
         tx = transport.send(0xe0, 0x04, 0x09, 0x00, data);
-        await sim.waitUntilScreenIsNot(snapshot2);
-        await sim.clickBoth(undefined, false);
+        await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
+        await handleKeyUi();
     } else if (type === CredentialDeploymentType.EXISTING) {
         data = Buffer.from('0120a845815bd43a1999e90fbf971537a70392eb38f89e6bd32b3dd70e1a9551d7', 'hex');
         tx = transport.send(0xe0, 0x04, 0x09, 0x00, data);
-        await sim.waitUntilScreenIsNot(snapshot2);
+        await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
         await handleAddressUi();
     }
 
@@ -92,20 +114,18 @@ async function credentialDeployment(
 async function updateCredentials(
     sim: Zemu, transport: Transport,
     handleInitialUi: () => Promise<any>,
-    handleKeyUi: () => Promise<any>,
     handleCredentialUi: () => Promise<any>,
 ) {
     const data = Buffer.from('08000004510000000000000000000000000000000000000002000000000000000020a845815bd43a1999e90fbf971537a70392eb38f89e6bd32b3dd70e1a9551d7000000000000000a0000000000000064000000290000000063de5da71402', 'hex');
     transport.send(0xe0, 0x31, 0x00, 0x00, data);
     await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
-    const snapshot = await handleInitialUi();
+    await handleInitialUi();
 
-    let credentialSnapshot;
     for (let i = 0; i < 2; i += 1) {
         const credentialIndex = Buffer.from(`0${i}`, 'hex');
         await transport.send(0xe0, 0x31, 0x00, 0x01, credentialIndex);
-        credentialSnapshot = await sharedCredentialDeployment(
-            snapshot, sim, transport, 0x31, 0x02, handleKeyUi,
+        await sharedCredentialDeployment(
+            sim, transport, 0x31, 0x02
         );
     }
 
@@ -115,6 +135,7 @@ async function updateCredentials(
 
     await transport.send(0xe0, 0x31, 0x00, 0x03, Buffer.from('02', 'hex'));
 
+    let credentialSnapshot = sim.getMainMenuSnapshot();
     for (const credentialId of credentialIdList) {
         transport.send(0xe0, 0x31, 0x00, 0x04, credentialId);
         await sim.waitUntilScreenIsNot(credentialSnapshot);
@@ -149,13 +170,6 @@ test('[NANO S] Update credentials', setupZemu('nanos', async (sim, transport) =>
             await sim.clickRight();
             await sim.clickRight();
             await sim.clickRight();
-            await sim.clickRight(undefined, false);
-            return sim.clickBoth(undefined, false);
-        },
-        async () => {
-            await sim.clickRight();
-            await sim.clickRight();
-            await sim.clickRight();
             await sim.clickRight();
             await sim.clickRight();
             await sim.clickRight(undefined, false);
@@ -176,10 +190,6 @@ async function updateCredentialsXAndSP(sim: Zemu, transport: Transport) {
         },
         async () => {
             await sim.clickRight();
-            return sim.clickBoth(undefined, false);
-        },
-        async () => {
-            await sim.clickRight();
             await sim.clickRight(undefined, false);
             return sim.clickBoth(undefined, false);
         },
@@ -189,6 +199,27 @@ async function updateCredentialsXAndSP(sim: Zemu, transport: Transport) {
 test('[NANO SP] Update credentials', setupZemu('nanosp', updateCredentialsXAndSP));
 
 test('[NANO X] Update credentials', setupZemu('nanox', updateCredentialsXAndSP));
+
+test('[NANO S] Credential deployment for new account with multiple keys', setupZemu('nanos', async (sim, transport) => {
+    await credentialDeployment(
+        sim,
+        transport,
+        CredentialDeploymentType.NEW,
+        'd68da82666d2e7eddbef3cb5cbeb95bfec62ee105e5f5a63d0db5fdc51934cfc6c0509316c06b384fa7df08eecafee9bbff378bc90ff436f26fa2ac8159ecd049000',
+        async () => {
+            await sim.navigate('.', 'multiple keys', [6, 0], true, false);
+        },
+        async () => {
+            // Not used for new account
+        },
+        3,
+        async () => {
+            await sim.navigate(".", 'multiple-keys', [4], true, false)
+            await sim.clickBoth(undefined, false);
+        },
+    );
+}));
+
 
 test('[NANO S] Credential deployment for new account', setupZemu('nanos', async (sim, transport) => {
     await credentialDeployment(
@@ -200,7 +231,10 @@ test('[NANO S] Credential deployment for new account', setupZemu('nanos', async 
             await sim.clickRight();
             await sim.clickRight();
             await sim.clickRight();
-            await sim.clickRight(undefined, false);
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
             return sim.clickBoth(undefined, false);
         },
         async () => {
@@ -218,7 +252,7 @@ test('[NANO S] Credential deployment for an existing account', setupZemu('nanos'
         async () => {
             await sim.clickRight();
             await sim.clickRight();
-            await sim.clickRight(undefined, false);
+            await sim.clickRight();
             return sim.clickBoth(undefined, false);
         },
         async () => {
@@ -226,7 +260,14 @@ test('[NANO S] Credential deployment for an existing account', setupZemu('nanos'
             await sim.clickRight();
             await sim.clickRight();
             await sim.clickRight();
-            await sim.clickRight(undefined, false);
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
             await sim.clickBoth(undefined, false);
         },
     );
@@ -239,7 +280,11 @@ async function credentialDeploymentNewAccountXAndSP(sim: Zemu, transport: Transp
         CredentialDeploymentType.NEW,
         '3a7d502e14ffd0e833bde629c798afc0721a6964af71f1c4dc54fdcde6ba0ce13db70f58bf9b3c2b9bd3fa100e1d6e21fe8b902237264c8662307bc8f7634f019000',
         async () => {
-            await sim.clickRight(undefined, false);
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
             return sim.clickBoth(undefined, false);
         },
         async () => {
@@ -264,6 +309,11 @@ async function credentialDeploymentExistingAccountXAndSP(sim: Zemu, transport: T
         },
         async () => {
             await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
             await sim.clickRight(undefined, false);
             await sim.clickBoth(undefined, false);
         },
@@ -273,3 +323,31 @@ async function credentialDeploymentExistingAccountXAndSP(sim: Zemu, transport: T
 test('[NANO SP] Credential deployment for an existing account', setupZemu('nanosp', credentialDeploymentExistingAccountXAndSP));
 
 test('[NANO X] Credential deployment for an existing account', setupZemu('nanox', credentialDeploymentExistingAccountXAndSP));
+
+async function credentialDeploymentNewAccountMultipleKeysXAndSP(sim: Zemu, transport: Transport) {
+    await credentialDeployment(
+        sim,
+        transport,
+        CredentialDeploymentType.NEW,
+        'd68da82666d2e7eddbef3cb5cbeb95bfec62ee105e5f5a63d0db5fdc51934cfc6c0509316c06b384fa7df08eecafee9bbff378bc90ff436f26fa2ac8159ecd049000',
+        async () => {
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            await sim.clickRight();
+            return sim.clickBoth();
+        },
+        async () => {
+            // Not used for new account
+        },
+        3,
+        async () => {
+            await sim.navigate(".", 'multiple-keys', [2], true, false)
+            await sim.clickBoth(undefined, false);
+        }
+    );
+}
+
+test('[NANO SP] Credential deployment for new account with multiple keys', setupZemu('nanosp', credentialDeploymentNewAccountMultipleKeysXAndSP));
+
+test('[NANO X] Credential deployment for new account with multiple keys', setupZemu('nanox', credentialDeploymentNewAccountMultipleKeysXAndSP));