+++ tests/src/options.ts
@@ -1,6 +1,7 @@
 import { resolve } from 'path';
 import Zemu, { DEFAULT_START_OPTIONS, IStartOptions } from '@zondax/zemu';
 import Transport from '@ledgerhq/hw-transport';
+import { ISnapshot } from '@zondax/zemu/dist/types';
 
 const SEED_PHRASE = 'vendor sphere crew wise puppy wise stand wait tissue boy fortune myself hamster intact window garment negative dynamic permit genre limb work dial guess';
 
@@ -31,26 +32,36 @@ export const optionsNanoSPlus: IStartOptions = {
 export const optionsNanoX: IStartOptions = {
     ...sharedOptions,
     model: 'nanox',
-    // set APILEVEL to 1, to ensure speculos uses a compatible SDK
-    custom: `${sharedOptions.custom}-a 1`,
+    // set APILEVEL to 5, to ensure speculos uses a compatible SDK
+    custom: `${sharedOptions.custom}-a 5`,
 };
 
 export const NANOS_ELF_PATH = resolve('bin/nanos/concordium_nanos.elf');
 export const NANOX_ELF_PATH = resolve('bin/nanox/concordium_nanox.elf');
 export const NANOS_PLUS_ELF_PATH = resolve('bin/nanosplus/concordium_nanosplus.elf');
 
-export function setupZemu(device: 'nanos' | 'nanosp' | 'nanox', func: (sim: Zemu, transport: Transport, device: 'nanos' | 'nanosp' | 'nanox') => Promise<void>) {
+class ConcordiumZemu extends Zemu {
+    // This is a hack to get around the issue where a screen is not fully rendered, but
+    // the screen has changed from the provided screenshot. This method waits a little bit
+    // after the screen has changed in an attempt to ensure that it is fully rendered.
+    async waitUntilScreenIsNot(screen: ISnapshot, timeout?: number | undefined): Promise<void> {
+        await super.waitUntilScreenIsNot(screen, timeout);
+        await Zemu.sleep(100);
+    }
+}
+
+export function setupZemu(device: 'nanos' | 'nanosp' | 'nanox', func: (sim: ConcordiumZemu, transport: Transport, device: 'nanos' | 'nanosp' | 'nanox') => Promise<void>) {
     return async () => {
         let sim;
         let simOptions;
         if (device === 'nanos') {
-            sim = new Zemu(NANOS_ELF_PATH);
+            sim = new ConcordiumZemu(NANOS_ELF_PATH);
             simOptions = optionsNanoS;
         } else if (device === 'nanosp') {
-            sim = new Zemu(NANOS_PLUS_ELF_PATH);
+            sim = new ConcordiumZemu(NANOS_PLUS_ELF_PATH);
             simOptions = optionsNanoSPlus;
         } else {
-            sim = new Zemu(NANOX_ELF_PATH);
+            sim = new ConcordiumZemu(NANOX_ELF_PATH);
             simOptions = optionsNanoX;
         }